二进制和json解包都使用dictionary，将代码共用。

+++ test/objc/mytest/mytest/ffan/dop/FFANDOPProtocol.h
@@ -40,6 +40,6 @@
 
 @property (nonatomic, retain)NSDictionary *sub_struct;
 
-@property (nonatomic, assign)int int_type;
+@property (nonatomic, assign)uint8_t int_type;
 
 @end
